requirements.txt
infi.systray
win10toast
flask
pycryptodomex
jsonpickle
click